+++ packages/types/custom/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add fields `hovered` and `selected` to core store.
 - Feature: Add parameters `renderType`, `featureList`, and `activeLayerPath` to `GfiConfiguration`.
 - Feature: Add type `FeatureList` for `GfiConfiguration`.
 - Feature: Add parameter `renderType` to `GeoLocationConfiguration`.